% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_long}
\alias{check_long}
\title{Check Long-Format Data File}
\usage{
check_long(
  long_data,
  timecol = NULL,
  groupcol = NULL,
  IDcol = NULL,
  measurementcol = NULL
)
}
\arguments{
\item{long_data}{The input data.frame in long format.}
}
\value{
The infile, checked and cleaned, still in long format.
}
\description{
This SQUEAK function checks a long-format data frame for downstream compatibility.
It will error out if problems are detected, and otherwise return an analysis-ready data frame.
It expects a data frame with the columns: 'time','group','ID','measurement'
The optional parameters (timecol etc) allow (and should be used) the user to specify the names of the columns that should be used instead of the above.
}
