% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{column_check}
\alias{column_check}
\title{Column Check for Long Format Data Frae}
\usage{
column_check(
  long_data,
  timecol = NULL,
  groupcol = NULL,
  IDcol = NULL,
  measurementcol = NULL
)
}
\arguments{
\item{long_data}{The input data.frame in long format.}

\item{timecol}{Name of the 'time' column (optional).}

\item{groupcol}{Name of the 'group' column (optional).}

\item{IDcol}{Name of the 'ID' column (optional).}

\item{measurementcol}{Name of the 'measurement' column (optional).}
}
\value{
The infile, checked and cleaned, still in long format.
}
\description{
This internal SQUEAK function looks at the columns of a data.frame in long format.
It will errors out if common issues are detected.
It expects a data frame with the columns: 'time','group','ID','measurement'
The optional parameters (timecol etc) allow the user to specify the names of the columns that should be used instead of the above. The analyses this package is geared for are essentially predicated on having: multiple measurements (for example, tumour volume) over time (days, weeks, etc), across multiple individuals (animal models) and treatment arms (or strains, etc).
}
