% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_mice}
\alias{long_mice}
\title{Tumour volume measurements from a mice model dataset.}
\format{
A data frame with 475 rows and 4 variables
}
\usage{
data(long_mice)
}
\description{
An example dataset containing longitudinal measures of  tumour volume across 4 drug arms.
}
\details{
\itemize{
\item Days the day that the measurement was taken
\item Group the drug arm each mouse was treated with
\item Number the ID of each individual mouse, unique to the drug arm
\item Value the measurement value, in this case tumour volume (cm3)
}
}
\keyword{datasets}
