% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_modelled_curves}
\alias{plot_modelled_curves}
\title{Plot Estimated Curves from the ME Model}
\usage{
plot_modelled_curves(
  model,
  xtitle = NULL,
  ytitle = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{xtitle}{A custom x-axis title for the graph (optional)}

\item{ytitle}{A custom y-axis title for the graph (optional)}

\item{title}{A custom title for the graph (optional)}

\item{subtitle}{A custom title for the graph (optional)}

\item{ready_data}{The mixed-effect model produced by the mixed_effect_model function}
}
\value{
A graph in ggplot format, plotting one growth curve per the grouping variable (eg. drug arm).
}
\description{
This SQUEAK function uses analysis-ready data and plots curves for each group
No y-axis transformation is applied, so curves should be apparent.
It performs minimal error checking and will crash ungracefully if the wrong things are given to it.
The optional parameters allow the user to specify axis/title labels.
}
\examples{
data("long_mice")
ready_data = check_long(long_mice,
                        timecol = "Days",
                        IDcol = 'Number',
                        groupcol = 'Group',
                        measurementcol = 'Value')
mixed_model = mixed_effect_model(ready_data = ready_data)
plot_modelled_curves(ready_data = ready_data ) # expects the columns: time, group, ID and measurement
}
