% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_raw_slopes}
\alias{plot_raw_slopes}
\title{Plot Raw Slopes Per Individual+Group}
\usage{
plot_raw_slopes(
  ready_data,
  title = NULL,
  subtitle = NULL,
  xtitle = NULL,
  ytitle = NULL
)
}
\arguments{
\item{long_data}{The input data.frame in long format.}
}
\value{
A nice faceted plot.
}
\description{
This SQUEAK function uses analysis-ready data and plots raw slopes (simple linear model), one per individual.
Log10 transformation is applied to the y-axis so make linear models usable.
Grouping variable is used to see different treatments etc.
It performs minimal error checking and will crash ungracefully if the wrong things are given to it.
The optional parameters allow the user to specify axis/title labels.
}
