% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{set_palette}
\alias{set_palette}
\title{Set a Palette for SQUEAK Plots}
\usage{
set_palette(ready_data, brewer_name = NULL)
}
\arguments{
\item{ready_data}{Analysis-ready data frame produced by check_long().}

\item{brewer_name}{What colorbrewer palette you'd like to draw from? (optional).}
}
\value{
The palette, a named vector.
}
\description{
This small function makes a palette for the SQUEAK plots.
It will errors out if common issues are detected.
It expects a data.frame created by the check_long function, with a column named 'group'.
In the future, optional parameters will work to let you change the base palette etc.
}
