% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUEAK-package.R
\docType{package}
\name{SQUEAK-package}
\alias{SQUEAK}
\alias{SQUEAK-package}
\title{SQUEAK: Statistical Quantification & Understanding of Experimental Animal Kinetics}
\description{
This package applies mixed-effect regressions to longitudinal studies of tumour growth data in mouse studies, with the goal of creating user-friendly output plots and some helpful statistics. Plots depend on the grammar of graphics (ggplot) ecosystem.
}
\author{
\strong{Maintainer}: George Seed \email{seedgeorge@gmail.com} (\href{https://orcid.org/0000-0002-1105-1781}{ORCID})

Authors:
\itemize{
  \item Jan Rekowski \email{jan.rekowski@icr.ac.uk}
}

}
\keyword{internal}
