% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_long}
\alias{check_long}
\title{Check Long-Format Data File}
\usage{
check_long(
  long_data,
  timecol = NULL,
  groupcol = NULL,
  IDcol = NULL,
  measurementcol = NULL,
  set_nested = TRUE
)
}
\arguments{
\item{long_data}{The input data.frame in long format.}

\item{timecol}{Name of the 'time' column (optional).}

\item{groupcol}{Name of the 'group' column (optional).}

\item{IDcol}{Name of the 'ID' column (optional).}

\item{measurementcol}{Name of the 'measurement' column (optional).}

\item{set_nested}{Individual IDs should usually be unique to treatment groups. If this is not desired, set this flag to FALSE.}
}
\value{
The infile, checked and cleaned, still in long format.
}
\description{
This SQUEAK function checks a long-format data frame for downstream compatibility.
It will error out if problems are detected, and otherwise return an analysis-ready data frame.
It expects a data frame with the columns: 'time','group','ID','measurement'
The optional parameters (timecol etc) allow the user to specify the names of the columns that should be used instead of the above. The analyses this package is geared for are essentially predicated on having: multiple measurements (for example, tumour volume) over time (days, weeks, etc), across multiple individuals (animal models) and treatment arms (or strains, etc).
}
