% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{mixed_effect_model}
\alias{mixed_effect_model}
\title{Run Mixed Effect Model}
\usage{
mixed_effect_model(ready_data, random_slope = TRUE, random_intercept = TRUE)
}
\arguments{
\item{ready_data}{The input data.frame in long format, containing the columns: time, group, ID and measurement (required)}

\item{random_slope}{Option to set a random slope for each individual in the model (optional)}

\item{random_intercept}{Option to set a random intercept for each individual in the model (optional)}
}
\value{
The model fit.
}
\description{
This SQUEAK function wraps some code from nlme to run a mixed-effect regression model.
It will error out ungracefully if problems are detected, and expects a data frame with the columns: 'time','group','ID','measurement' It is suggested to run the utility function check_long() to determine if your data is the correct format.
}
