% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_modelled_curves}
\alias{plot_modelled_curves}
\title{Plot Estimated Curves from the ME Model}
\usage{
plot_modelled_curves(
  model,
  xtitle = NULL,
  ytitle = NULL,
  title = NULL,
  subtitle = NULL,
  palette = NULL,
  grouptitle = NULL,
  ylim = NULL,
  show_confidence = TRUE
)
}
\arguments{
\item{model}{The mixed-effect model produced by the mixed_effect_model function}

\item{xtitle}{A custom x-axis title for the graph (optional)}

\item{ytitle}{A custom y-axis title for the graph (optional)}

\item{title}{A custom title for the graph (optional)}

\item{subtitle}{A custom title for the graph (optional)}

\item{palette}{A custom palette for the graph, from the set_palette() function (optional)}

\item{grouptitle}{A custom title for the grouping variable, replacing the default 'group' (optional)}

\item{ylim}{A custom set of y-axis limits, as a two-element vector ie. c(0,100) (optional)}

\item{show_confidence}{Do you want to see confidence intervals? TRUE/FALSE, default=TRUE (optional)}
}
\value{
A graph in ggplot format, plotting one growth curve per the grouping variable (eg. drug arm).
}
\description{
This SQUEAK function uses analysis-ready data and plots curves for each group
No y-axis transformation is applied, so curves should be apparent.
It performs minimal error checking and will crash ungracefully if the wrong things are given to it.
The optional parameters allow the user to specify axis/title labels.
}
\examples{
data("long_mice")
ready_data = check_long(long_mice,
                        timecol = "Days",
                        IDcol = 'Number',
                        groupcol = 'Group',
                        measurementcol = 'Value')
mixed_model = mixed_effect_model(ready_data = ready_data)
plot_modelled_curves(model = mixed_model ) 
}
